requirements.txt
aioredis==1.3.1
ansicon==1.89.0
arabic-reshaper==3.0.0
arrow==1.2.3
asgiref==3.10.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
babel==2.17.0
blessed==1.19.1
certifi==2022.9.24
cffi==2.0.0
channels==2.4.0
channels-redis==2.4.2
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==46.0.3
cssselect2==0.8.0
daphne==2.5.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.8
django-allauth==0.63.6
django-autoslug==1.9.9
django-babel==0.6.2
django-braces==1.17.0
django-cacheops==7.2
django-cities-light==3.10.1
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-cors-headers==4.5.0
django-countries==7.4.2
django-dotenv==1.4.2
django-extensions==3.2.3
django-filter==23.5
django-js-asset==2.2.0
django-model-utils==5.0.0
django-password-reset==2.0
django-picklefield==3.2
django-private-chat==0.3.0
django-q==1.3.9
django-ranged-response==0.2.0
django-redis==5.4.0
django-simple-captcha==0.5.20
django-utils==0.0.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
easy-thumbnails==2.8.3
ecdsa==0.18.0
freetype-py==2.5.1
funcy==1.17
gunicorn==20.1.0
hiredis==3.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jinxed==1.2.0
legacy-cgi==2.6.2
lxml==6.0.2
Markdown==3.4.1
msgpack==0.6.2
mysqlclient==2.2.7
oauth2_provider==0.0
oauth_provider==1.1
oauthlib==3.2.2
oscrypto==1.3.0
packaging==21.3
pillow==11.1.0
progressbar2==4.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.29.0
pycparser==2.21
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.6.0
pyOpenSSL==25.3.0
pyparsing==3.0.9
pypdf==6.5.0
python-bidi==0.6.7
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-utils==3.9.1
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0.3
redis==3.5.3
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==1.3.1
rlPyCairo==0.4.0
rsa==4.9
service-identity==21.1.0
setuptools==75.8.2
six==1.16.0
sqlparse==0.4.3
svglib==1.6.0
tinycss2==1.5.1
Twisted==24.3.0
txaio==22.2.1
typing_extensions==4.12.2
tzdata==2022.6
tzlocal==5.3.1
Unidecode==1.3.8
uritools==6.0.1
urllib3==1.26.12
waitress==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
wheel==0.45.1
wrapt==1.14.1
xhtml2pdf==0.2.17
zope.interface==7.2
